+++ frontend/src/components/form/QueueForm.tsx
@@ -3,28 +3,14 @@ import { useDispatch, useSelector } from "react-redux";
 import React from "react";
 import {
   enterQueue,
-  toggleDifficulty,
+  toggleDifficulty
 } from "../../feature/matching/matchingSlice";
 import { JoinQueueRequest } from "../../proto/matching-service";
 import { RootState } from "../../app/store";
 import useFixedToast from "../../utils/hooks/useFixedToast";
 import { QuestionDifficulty } from "../../proto/types";
 import MatchingAPI from "../../api/matching";
-
-const DIFFICULTY = [
-  {
-    name: "Easy",
-    colorScheme: "green",
-  },
-  {
-    name: "Medium",
-    colorScheme: "orange",
-  },
-  {
-    name: "Hard",
-    colorScheme: "red",
-  },
-];
+import { DIFFICULTY } from "../../constants/difficulty";
 
 function QueueForm() {
   const dispatch = useDispatch();
@@ -39,7 +25,6 @@ function QueueForm() {
   };
 
   const enterQueueHandler = () => {
-    // API call to enter queue, probably may need to pass some information to redux store
     if (noneSelected) {
       toast.sendErrorMessage("You must select at least 1 difficulty");
       return;
@@ -52,7 +37,7 @@ function QueueForm() {
       .forEach((x) => selectedDifficulties.push(x as QuestionDifficulty));
 
     const joinQueueReq: JoinQueueRequest = {
-      difficulties: selectedDifficulties,
+      difficulties: selectedDifficulties
     };
 
     MatchingAPI.joinQueue(joinQueueReq)
