+++ frontend/src/constants/difficulty.ts
@@ -1,0 +1,17 @@
+const DIFFICULTY = [
+	{
+		name: "Easy",
+		colorScheme: "green",
+	},
+	{
+		name: "Medium",
+		colorScheme: "orange",
+	},
+	{
+		name: "Hard",
+		colorScheme: "red",
+	},
+];
+
+// eslint-disable-next-line
+export { DIFFICULTY };
